add controller
