+++ models/contacts.json
@@ -58,11 +58,5 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
-  },
-  {
-    "id": "WMB29yjuiLHF7JHjYpI0h",
-    "name": "Nata",
-    "email": "da2012@gmail.com",
-    "phone": "000443792"
   }
 ]
